% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AGBChv05DH}
\alias{AGBChv05DH}
\title{Function to estimate individual biomass for dry forests (Chave, 2005).}
\usage{
AGBChv05DH(xdataset, dbh = "DBH4")
}
\arguments{
\item{xdataset}{a dataset for estimating biomass}

\item{dbh}{a diameter (in mm).}
}
\description{
Function to estimate individual biomass using Chave (2005) equation for dry forests. Height is estimated using Feldpaush (2011) regional parameters with Weibull equation. The function adds columns to the dataset with the biomass information for all alive trees. This function needs a dataset with the following information: PlotViewID, PlotID, TreeID, CensusNo, Diameter (DBH1-DBH4), Wood density (WD) and Allometric RegionID.The function assumes that the diameter used is DBH4, unless other DBH is selected. See ForestPlots.net documentation for more information.
}
\author{
Gabriela Lopez-Gonzalez
}
\references{
Chave C, Andalo S, Brown, et al. 2005. Tree allometry and improved estimation of carbon stocks and balance in tropical forests. Oecologia 145 (1):87-99. doi:10.1007/s00442-005-0100-x.

Feldpausch TR, Banin L, Phillips OL, Baker TR, Lewis SL et al. 2011. Height-diameter allometry of tropical forest trees. Biogeosciences 8 (5):1081-1106. doi:10.5194/bg-8-1081-2011.

Chave J, Coomes DA, Jansen S, Lewis SL, Swenson NG, Zanne AE. 2009. Towards a worldwide wood economics spectrum. Ecology Letters 12(4): 351-366. http://dx.doi.org/10.1111/j.1461-0248.2009.01285.x

Zanne AE, Lopez-Gonzalez G, Coomes DA et al. 2009. Data from: Towards a worldwide wood economics spectrum. Dryad Digital Repository. http://dx.doi.org/10.5061/dryad.234
}

