\name{BiomasaFP-package}
\alias{BiomasaFP-package}
\alias{BiomasaFP}
\docType{package}
\title{
ForestPlots.net Package to estimate aboveground biomass (AGB) and rates of AGB change.

}
\description{
 Tools for analysing data downloaded from ForestPlots.net. This
    package includes functions for: a) estimating above ground biomass (AGB)
    from forest inventory data and b) producing summary statistics for each plot and census. 
}
\details{
\tabular{ll}{
Package: \tab BiomasaFP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-03\cr
License: \tab GPL-3\cr
}
To use this package you need to download the following data sets from ForestPlots.net: 
a) Individuals dataset from the Advanced search;
b) Individual Wood density values from the Query Library
c) Plot Metadata for R, from the Query Library.

}
\author{
Gabriela Lopez-Gonzalez, Martin Sullivan, Tim Baker

Maintainer: Gabriela Lopez-Gonzalez <admin@forestplots.net>

}
\references{
Lopez-Gonzalez G, Lewis SL, Burkitt M. and Phillips OL. 2011. ForestPlots.net: a web application and research tool to manage and analyse tropical forest plot data. Journal of Vegetation Science 22: 610-613. doi: 10.1111/j.1654-1103.2011.01312.x

Biomass and height estimates

Chave C, Andalo S, Brown, et al. 2005. Tree allometry and improved estimation of carbon stocks and balance in tropical forests. Oecologia 145 (1):87-99. doi:10.1007/s00442-005-0100-x.

Chave J, Rejou-Mechain M, Burquez A et al. 2014. Improved allometric models to estimate the aboveground biomass of tropical trees. Global Change Biology 20: 3177-3190. doi: 10.1111/gcb.12629

Feldpausch TR, Banin L, Phillips OL, Baker TR, Lewis SL et al. 2011. Height-diameter allometry of tropical forest trees. Biogeosciences 8 (5):1081-1106. doi:10.5194/bg-8-1081-2011.

Wood Density

Chave J, Coomes DA, Jansen S, Lewis SL, Swenson NG, Zanne AE. 2009. Towards a worldwide wood economics spectrum. Ecology Letters 12(4): 351-366. http://dx.doi.org/10.1111/j.1461-0248.2009.01285.x

Zanne AE, Lopez-Gonzalez G, Coomes DA, Ilic J, Jansen S, Lewis SL, Miller RB, Swenson NG, Wiemann MC, Chave J. 2009. Data from: Towards a worldwide wood economics spectrum. Dryad Digital Repository. http://dx.doi.org/10.5061/dryad.234



}

