% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weib_models.r
\name{fit.weib.models}
\alias{fit.weib.models}
\title{fit.weib.models}
\usage{
fit.weib.models(x, thresh = 20, weib.only = FALSE)
}
\arguments{
\item{x}{Dataframe with columns names Height (consisting of measured heights), Diameter (measured diameters) and WD (wood density). These will be present if an object return by \code{mergefp} is used.}

\item{thresh}{Minimum number of height observations needed to attempt to fit models}

\item{weib.only}{Logical. Should the search for the best model be constrained to Weibull models. If TRUE, log-log models are only selected when neither Weibull models converge. Default is FALSE.}
}
\value{
Dataframe with estimated parameters, number of height measurments used to fit the models and the error at estimating AGB compared to AGB estimates using observed heights (in units of Mg per tree). Parameters of the 'best' model (i.e. lowest AGB prediction error) are also given.
}
\description{
Function to fit Weibull and log-log height-diameter models.
}
\author{
Martin Sullivan
}

