% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_weib.r
\name{fit.weib}
\alias{fit.weib}
\title{Function to fit height-diamter models}
\usage{
fit.weib(data, return.mods = FALSE)
}
\arguments{
\item{data}{Object returned by mergefp}

\item{return.mods}{Logical. If TRUE, full model fits are returned. Otherwise (default) only the fitted coefficents are returned.}
}
\value{
A data frame with PlotViewID and fitted parameters of Weibull models at each level.
}
\description{
Function for fitting height-diameter Weibull models by non-linear least squares
}
\author{
Martin Sullivan, Gabriela Lopez-Gonzalez
}

