% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodmanPalm.R
\name{GoodmanPalm}
\alias{GoodmanPalm}
\title{GoodmanPalm}
\usage{
GoodmanPalm(data, dbh = "D4")
}
\arguments{
\item{data}{Object returned by \code{mergefp}.}

\item{dbh}{Name of column containing diameter data. Default is "D4".}
}
\value{
Vector with estimated AGB for each row in /code{data}.
}
\description{
Function to estimate AGB of palms using family level diameter based equation from Goodman et al. 2013.
}
\author{
Martin Sullivan
}

