% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_fit.r
\name{hd.fit}
\alias{hd.fit}
\title{hd.fit}
\usage{
hd.fit(htdata, level, weib.only = FALSE, thresh = 20, forest.type = TRUE)
}
\arguments{
\item{htdata}{Dataframe containing height measurements.}

\item{level}{Level to fit height-diameter model. One of "Plot", "Cluster", "BioR" and "Continent".}

\item{weib.only}{Logical. Should the search for the best model be constrained to Weibull models. If TRUE, log-log models are only selected when neither Weibull models converge. Default is FALSE.}

\item{thresh}{thresh argument to be passed to /code{fit.weib.models}}

\item{forest.type}{Logical. If true (default), level is constrained to be within a particular forest type/}
}
\value{
Dataframe with estimated parameters, number of height measurments used to fit the models and the error at estimating AGB compared to AGB estimates using observed heights (in units of Mg per tree). Parameters of the 'best' model (i.e. lowest AGB prediction error) are also given.
}
\description{
Fit height-diameter models for each level in a dataset. This is a wrapper for \code{fit.weib.models}.
}
\author{
Martin Sullivan
}

