% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_constrain.r
\name{height.constrain}
\alias{height.constrain}
\title{height.constrain}
\usage{
height.constrain(fit, data, level, dbh = "D4", max.size = 80,
  na.too.big = FALSE, forest.type = TRUE)
}
\arguments{
\item{fit}{Fitted height-diameter models object returned by /code{hd.fit}.}

\item{data}{Object returned by /code{mergefp}.}

\item{level}{Level to fit height-diameter model. One of "Plot", "Cluster", "BioR" and "Continent". Must match level used whe constructing fit object.}

\item{dbh}{Name of column containing diameter measurements. Default is "D4".}

\item{max.size}{Maximum tree height. If all models predict the height of the largest tree in the plot to be below this, then the model with the lowest AGB prediction error will be used. Otherwise models predicting heights above this will be excluded from selection. If all models predict heights above this, then the model giving the lowest maximum height will be selected.}

\item{na.too.big}{Logical. Should parameters be set to NA if all models exceed height treshold. Default is FALSE.}

\item{forest.type}{Logical. Should ID include forest type.}
}
\value{
Dataframe with components returned by /code{hd.fit}, but with addtional columns giving the maximum diameter of the largest tree in the plot, the predicted maximum height using each model, and a column indicating whether the choice of best model was modified (0 = no change, 1 = previous best model excluded due to exceeding threshold, 2 = all models exceeded height threshold).
}
\description{
Modify selection of fitted height-diameter models to avoid models that overpredict over a threshold maximum height
}
\author{
Martin Sullivan
}

