% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_mod.r
\name{height.mod}
\alias{height.mod}
\title{height.mod}
\usage{
height.mod(dbh, a, b, c = NA)
}
\arguments{
\item{dbh}{Vector of diameters.}

\item{a}{a parameter in Weibull or log-log model}

\item{b}{b parameter in Weibull or log-log model}

\item{c}{c parameter in Weibull model. If NA and other parameters are not, the function assumes the model is a log-log model}
}
\value{
Vector with estimated heights.
}
\description{
Function to estimate height from height-diameter model parameters
}
\author{
Martin Sullivan
}

