% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_heights.r
\name{local.heights}
\alias{local.heights}
\title{local.heights}
\usage{
local.heights(xdataset, dbh = "D4", use.height.constrain = FALSE,
  max.size = 80, thresh = 20, na.too.big = FALSE, laser.cor = FALSE,
  no.plot = TRUE)
}
\arguments{
\item{xdataset}{Object returned by /code{mergefp}.}

\item{dbh}{Character giving the name of the column contining the diameter measurements. Defaults to "D4".}

\item{use.height.constrain}{Logical. Should height constrain function be used when selecting best model. Defaul is FALSE.}

\item{max.size}{max.size argument to /code{height.constrain}.}

\item{thresh}{thresh argument to /code{fit.weib.models}.}

\item{na.too.big}{na.too.big argumet to /code{height.constrain}.}

\item{laser.cor}{Logical. Should laser heights be adjusted to correct for bias. Default is FALSE.}

\item{no.plot}{Logial. Should the best parameters exclude plot and cluster level. Default is TRUE.}
}
\value{
Dataframe with best parameters for height-diameter models for each row in xdataset.
}
\description{
Wrapper function for estimating heights using local allometric equations.
}
\author{
Martin Sullivan
}

