% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SizeClassGrowth.r
\name{SizeClassGrowth}
\alias{SizeClassGrowth}
\title{SizeClassGrowth}
\usage{
SizeClassGrowth(xdataset, method = "median", dbh = "D4")
}
\arguments{
\item{xdataset}{Object returned by \code{mergefp}}

\item{method}{Either "median" (default) or "mean". If any value other than "median" is entered the mean will be calculated.}

\item{dbh}{Name of column with diameter measurement to use. Default to "D4".}
}
\value{
A data frame with PlotViewID, medain/mean growth rates for each size class and mean wood density in each plot
}
\description{
Estimate mean or median growth rate for size classes 100-200mm DBH, 200-400mm DBH and >400mm DBH.
}
\author{
Martin Sullivan
}

