% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryAGWP.r
\name{SummaryAGWP}
\alias{SummaryAGWP}
\title{SummaryAGWP}
\usage{
SummaryAGWP(xdataset, AGBEquation, dbh = "D4", rec.meth = 0,
  height.data = NULL, param.type = "Best", palm.eq = TRUE)
}
\arguments{
\item{xdataset}{Object returned by \code{mergefp}.}

\item{AGBEquation}{Allometric equation function to use when estimating AGB.Note only \code{AGBChv14} and \code{AGBChv05MH} are fully implemented, other Chave 2005 equations can by used for observed components by unobserved components will use \code{AGBChv05MH}. Plan to implement other equations fully in future.}

\item{dbh}{Name of column containing diameter data. Default is "D4".}

\item{rec.meth}{Method used to estimate AGWP of recruits. If 0 (default), estimates growth from starting diameter of 0mm. If another value is provided, then growth is estimated from a starting diameter of 100mm.}

\item{height.data}{Object returned by \code{local.heights}. Used to supply parameters of local allometric equations. If NULL (default), regional height diameter equations are used.}

\item{param.type}{Local height diameter to use. One of "Best" (defualt), "ClusterF","BioRF" or "ContinentF". Ignored if \code{height.data=NULL}.}

\item{palm.eq}{Logical. If TRUE the family level diameter based equation from Goodman et al 2013 is used to estimate AGB of monocots. Otherwise AGB of monocots is estimated using the allometric equation supplied to \code{AGBEquation}.}
}
\value{
A data frame with PlotViewID, CensusNo, and observed and unobserved elements of AGWP, stem dynamics and AGB mortality.
}
\description{
Function to estimate AGWP by plotview and census interval, including estimating unobserved recruitment and growth of trees that died between census periods
}
\author{
Martin Sullivan, Gabriela Lopez-Gonzalez
}

