% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SummaryFamilies}
\alias{SummaryFamilies}
\title{Function for summarizing tree families by Plot}
\usage{
SummaryFamilies(xdataset)
}
\arguments{
\item{xdataset}{a dataset  that has been merged using the function mergefp.}
}
\description{
Function to summarize information on tree families (APGIII) recorded by Plot (PlotViewID).
Output headers are: Country, PlotCode, PlotViewID, Family, No.sp (number of species), Frequency (number of individuals in a family).
}
\author{
Gabriela Lopez-Gonzalez
}

