% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSummary.R
\name{SummaryPlotviews}
\alias{SummaryPlotviews}
\title{Function for summarizing Plot View metadata}
\usage{
SummaryPlotviews(xdataset)
}
\arguments{
\item{xdataset}{a dataset  that has been merged using the function mergefp.}
}
\description{
Function for summarizing the information of the Plot Views used (PlotViewID). Output headers are: Country, PlotCode, PlotViewID, LatitudeDecimal, LongitudeDecimal,
First.Census.Date (Decimal date when the Plot was established), No.Censuses (Number of censuses or field campaigns).
}
\author{
Gabriela Lopez-Gonzalez
}

