% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeParameters.R
\name{TreeParameters}
\alias{TreeParameters}
\title{Function to fit tree-level allometries.}
\usage{
TreeParameters(xdataset)
}
\arguments{
\item{xdataset}{Object returned by /code{local.heights}}
}
\description{
This function solves the a parameter in Weibull height-diameter models to force the model through a measured tree height
}
\author{
Wannes Hubau
}

