% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_merge.r
\name{param.merge}
\alias{param.merge}
\title{Merge fitted height-diameter parameters with tree by tree data}
\usage{
param.merge(data, wparm)
}
\arguments{
\item{data}{Object returned by mergefp}

\item{wparm}{Object returned by \\coding {fit.weib}}
}
\description{
Function to merge fitted parameters of Weibull height-diameter equations to tree data returned by \code{mergefp}
}
\author{
Martin Sullivan, Gabriela Lopez-Gonzalez
}

