% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummaryAGWP.r
\name{SummaryAGWP}
\alias{SummaryAGWP}
\title{SummaryAGWP}
\usage{
SummaryAGWP(xdataset, AGBEquation, dbh = "D4", rec.meth = 0,
  height.data = NULL, param.type = "Best")
}
\arguments{
\item{xdataset}{Object returned by \code{mergefp}.}

\item{AGBEquation}{Allometric equation function to use when estimating AGB.Note only \code{AGBChv14} and \code{AGBChv05MH} are fully implemented, other Chave 2005 equations can by used for observed components by unobserved components will use \code{AGBChv05MH}. Plan to implement other equations fully in future.}

\item{dbh}{Name of column containing diameter data. Default is "D4".}

\item{rec.meth}{Method used to estimate AGWP of recruits. If 0 (default), estimates growth from starting diameter of 0mm. If another value is provided, then growth is estimated from a starting diameter of 100mm.}

\item{height.data}{Object returned by \code{param.merge}. Used to supply parameters of local allometric equations. If NULL (default), regional height diameter equations are used.}

\item{param.type}{Local height diameter to use. One of Best (defualt), BioRF,ClusterF ... NEED TO DECIDE WHICH OF THESE TO RETURN. Ignored if \code{height.data=NULL}.}
}
\value{
A data frame with PlotViewID, CensusNo, and observed and unobserved elements of AGWP, stem dynamics and AGB mortality.
}
\description{
Function to estimate AGWP by plotview and census interval, including estimating unobserved recruitment and growth of trees that died between census periods
}
\author{
Martin Sullivan, Gabriela Lopez-Gonzalez
}

