% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FamilySummary.R
\name{SummaryFamilies}
\alias{SummaryFamilies}
\title{Function for summarizing tree families by plot}
\usage{
SummaryFamilies(xdataset)
}
\arguments{
\item{xdataset}{a dataset  that has been merged using the function mergefp.}
}
\description{
Function to summarize information on tree families (based on APGIII classification) by Plot.
Output headers are: Country, PlotCode, PlotViewID, Family, No.sp (number of species), Frequency (number of individuals in a family).
}
\author{
Gabriela Lopez-Gonzalez
}

